+++ cmd/oauth/oauth.go
@@ -25,6 +25,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/redhat-appstudio/service-provider-integration-operator/cmd"
 	"github.com/redhat-appstudio/service-provider-integration-operator/cmd/oauth/cli"
 	"github.com/redhat-appstudio/service-provider-integration-operator/oauth/metrics"
 
@@ -39,7 +40,6 @@ import (
 	oauth2 "github.com/redhat-appstudio/service-provider-integration-operator/pkg/serviceprovider/oauth"
 	"github.com/redhat-appstudio/service-provider-integration-operator/pkg/spi-shared/config"
 	"github.com/redhat-appstudio/service-provider-integration-operator/pkg/spi-shared/tokenstorage"
-	"github.com/redhat-appstudio/service-provider-integration-operator/pkg/spi-shared/tokenstorage/vaultstorage"
 	authz "k8s.io/api/authorization/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/meta"
@@ -99,11 +99,9 @@ func main() {
 		os.Exit(1)
 	}
 
-	vaultConfig := vaultstorage.VaultStorageConfigFromCliArgs(&args.VaultCliArgs)
-	vaultConfig.MetricsRegisterer = metrics.Registry
-	strg, err := vaultstorage.NewVaultStorage(vaultConfig)
+	strg, err := cmd.InitTokenStorage(context.Background(), &args.CommonCliArgs)
 	if err != nil {
-		setupLog.Error(err, "failed to create token storage interface")
+		setupLog.Error(err, "failed to initialize the token storage")
 		os.Exit(1)
 	}
