+++ pkg/spi-shared/tokenstorage/vaultstorage/vaultcli/cli_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package vaultstorage
+package vaultcli
 
 import (
 	"testing"
